Correct some mistake
1.Change CMAKE_CXX_FLAGS to CMAKE_CP_FLAGS
2.Change type of len in static functions of Num from int to size_t
3.And suffix u for all unsigned numbers

+++ Numer/Num.cpp
@@ -8,7 +8,7 @@ Num::Num(const Num&)
 {
 }
 
-uint32_t * Num::bitPatAdd(uint32_t* const opa, uint32_t* const opb, int len)
+uint32_t * Num::bitPatAdd(uint32_t* const opa, uint32_t* const opb, size_t len)
 {
 	uint32_t* rst = new uint32_t[len];
 	uint32_t c = 0u; // The carry
@@ -34,7 +34,7 @@ uint32_t * Num::bitPatAdd(uint32_t* const opa, uint32_t* const opb, int len)
 	return rst;
 }
 
-uint32_t * Num::bitPatSub(uint32_t* const opa, uint32_t* const opb, int len)
+uint32_t * Num::bitPatSub(uint32_t* const opa, uint32_t* const opb, size_t len)
 {
 	// a-b means a+(-b) in complement form
 	uint32_t* opbc = new uint32_t[len];
@@ -52,7 +52,7 @@ uint32_t * Num::bitPatSub(uint32_t* const opa, uint32_t* const opb, int len)
 	return Num::bitPatAdd(opa, opb, len);
 }
 
-bool Num::bitPatCompare(uint32_t * opa, uint32_t * opb, int len)
+bool Num::bitPatCompare(uint32_t * opa, uint32_t * opb, size_t len)
 {
 	for (size_t i = len - 1; i >= 0; i++)
 	{
